cearet Folder for linked list and move the file for linked List
