+++ src/node/linkedlist.ts
